Update class-divi-accessibility-public.php
PHP Warning:  Undefined array key "hide_aria_element" in /home/indept7/public_html/wp-content/plugins/divi-accessibility-master/public/class-divi-accessibility-public.php on line 408

This updated code snippet first checks if the hide_aria_element key is set in the $element->props array. If it exists and its value is equal to 'on', then it proceeds to append ' aria-hidden' to the $class_list. This approach will prevent the warning when the key does not exist in the array.

PHP Warning:  Undefined array key "show_for_screen_readers_only" in /home/linkcorp/public_html/wp-content/plugins/divi-accessibility-master/public/class-divi-accessibility-public.php on line 411

This modification first checks if the show_for_screen_readers_only key exists in the $element->props array. If it does, and its value is equal to 'on', then it will add ' screen-reader-text' to $class_list. This prevents the warning from occurring when the key is not present in the array.